+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
 
-## [Unreleased]
-
-
+## [0.1.0] - 2021-12-31
+New Repo For Kommunicate iOS Chat UI 
+- Added Support for Chat Custom Event Listeners
+- Added support for text areas in rich message forms
 - Send Post Submitted Form Data As Message
 - Added Support for RTL Languages